% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_windrose.R
\name{coord_windrose}
\alias{coord_windrose}
\title{Polar coordinate system for wind roses}
\usage{
coord_windrose(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:coord_radial]{ggplot2::coord_polar()}}.}
}
\value{
A \code{CoordPolar} object.
}
\description{
Wraps \code{\link[ggplot2:coord_radial]{ggplot2::coord_polar()}} with the following defaults:
\itemize{
\item \code{theta = "x"}: wind direction on the angular axis
\item \code{start = 0}: North (x = 0°) at 12 o'clock
\item \code{direction = 1}: clockwise (meteorological convention)
}
}
\details{
\tabular{lll}{
   Degrees \tab Compass \tab Position \cr
   0 / 360 \tab N \tab 12 o'clock \cr
   90 \tab E \tab 3 o'clock \cr
   180 \tab S \tab 6 o'clock \cr
   270 \tab W \tab 9 o'clock \cr
}
}
\examples{
# coord_windrose()

}
