% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides.R
\docType{data}
\name{GuideWindrose}
\alias{GuideWindrose}
\title{Guide ggproto object for wind rose legends}
\description{
The \code{Guide} ggproto class underlying \code{\link[=guide_windrose]{guide_windrose()}}.
Extends \code{GuideLegend} to render a calm-wind percentage footer.
Not intended for direct use.
}
\keyword{datasets}
