% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_windrose.R
\docType{data}
\name{StatWindrose}
\alias{StatWindrose}
\title{Stat ggproto object for wind roses}
\description{
The \code{Stat} ggproto class underlying \code{\link[=stat_windrose]{stat_windrose()}}. Bins
wind direction and speed into a direction-by-speed frequency table.
Not intended for direct use.
}
\keyword{datasets}
